Align OpprettModal for SM and AG
